Change chainId type to Long,fix evm private key import bugs below Android 8

+++ app/src/main/java/com/connect/demo/model/RpcRequest.kt
@@ -4,7 +4,7 @@ package com.connect.demo.model
  * Created by chaichuanfa on 2022/7/27
  */
 data class RpcRequest(
-    val chainId: Int,
+    val chainId: Long,
     val id: String,
     val jsonrpc: String = "2.0",
     val method: String,
